+++ user_frontend/src/App.tsx
@@ -1,13 +1,24 @@
-import Chatbox from "./components/chatbox";
-import Header from "./components/header";
+import { SignedIn, SignedOut } from "@clerk/clerk-react";
+import Chatbox from "./components/ChatBox";
+import Header from "./components/Header";
 
 function App() {
   return (
     <>
-      <Header />
-      <Chatbox />
+      <SignedIn>
+        <>
+          <Header mode="authenticated" />
+          <Chatbox mode="authenticated" />
+        </>
+      </SignedIn>
+      <SignedOut>
+        <>
+          <Header mode="public" />
+          <Chatbox mode="public" />
+        </>
+      </SignedOut>
     </>
-  )
+  );
 }
 
-export default App
+export default App;
